% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLayer.R
\name{add_layer_to_tree}
\alias{add_layer_to_tree}
\title{Internal function}
\usage{
add_layer_to_tree(tree)
}
\arguments{
\item{tree}{A tree data frame from \code{\link[=getTreeranger]{getTreeranger()}}.}
}
\value{
A tree data frame with \code{layer} added.
}
\description{
This function adds the respective layer to the different nodes in a tree.
The tree has to be prepared by getTree function.
}
\seealso{
\code{\link[=addLayer]{addLayer()}}
}
\keyword{internal}
