% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLayer.R
\name{addLayer}
\alias{addLayer}
\title{Add layer information to a forest that was created by getTreeranger}
\usage{
addLayer(trees, num.threads = 1)
}
\arguments{
\item{trees}{The output of \code{\link[=getTreeranger]{getTreeranger()}}.}

\item{num.threads}{(Default: 1) Number of threads to spawn for parallelization.}
}
\value{
A list of tree data frames of length \code{RF$num.trees}.
Each row of the tree data frames corresponds to a node of the respective tree and the columns correspond to:
\itemize{
\item \code{nodeID}: ID of the respective node (important for left and right daughters in the next columns)
\item \code{leftdaughter}: ID of the left daughter of this node
\item \code{rightdaughter}: ID of the right daughter of this node
\item \code{splitvariable}: ID of the split variable
\item \code{splitpoint}: Split point of the split variable.
For categorical variables this is a comma separated lists of values, representing the factor levels (in the original order) going to the right.
\item \code{status}: \code{0} for terminal (\code{splitpoint} is \code{NA}) and \code{1} for non-terminal.
\item \code{layer}: Tree layer depth information, starting at 0 (root node) and incremented for each layer.
}
}
\description{
This functions adds the layer information to each node in a list with trees that was obtained by getTreeranger.
You should use \code{\link[=getTreeranger]{getTreeranger()}} with \code{add_layer = TRUE} instead.
}
