% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSurrogates.R
\name{addSurrogates}
\alias{addSurrogates}
\title{Add surrogate information to a tree list.}
\usage{
addSurrogates(
  RF,
  trees,
  s,
  Xdata,
  num.threads = parallel::detectCores(),
  preschedule.threads = TRUE
)
}
\arguments{
\item{RF}{A [ranger::ranger] object which was created with `keep.inbag = TRUE`.}

\item{trees}{List of trees created by [getTreeranger].}

\item{s}{Predefined number of surrogate splits (it may happen that the actual number of surrogate splits differs in individual nodes).}

\item{Xdata}{data without the dependent variable.}

\item{num.threads}{(Default: [parallel::detectCores()]) Number of threads to spawn for parallelization.}

\item{preschedule.threads}{(Default: TRUE) Passed as `mc.preschedule` to [parallel::mclapply()].}
}
\value{
A list of trees.
A list of trees containing of lists of nodes with the elements:
* `nodeID`: ID of the respective node (important for left and right daughters in the next columns)
* `leftdaughter`: ID of the left daughter of this node
* `rightdaughter`: ID of the right daughter of this node
* `splitvariable`: ID of the split variable
* `splitpoint`: splitpoint of the split variable
* `status`: `0` for terminal and `1` for non-terminal
* `layer`: layer information (`0` means root node, `1` means 1 layer below root, etc)
* `surrogate_i`: numbered surrogate variables (number depending on s)
* `adj_i`: adjusted agreement of variable i
}
\description{
This function adds surrogate variables and adjusted agreement values to a forest that was created by [getTreeranger].
}
