% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSurrogates.R
\name{getSurrgate2}
\alias{getSurrgate2}
\title{getSurrogate2}
\usage{
getSurrgate2(x, Xdata, controls, s, ncat)
}
\arguments{
\item{x}{List of length `num.trees`}
}
\description{
This is an internal function
}
\keyword{internal}
