% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutualForestImpact.R
\name{MFI_VarSel_Permutation}
\alias{MFI_VarSel_Permutation}
\title{Permutation method for MFI.}
\usage{
MFI_VarSel_Permutation(MFI, candidates, variables, p.threshold = 0.01)
}
\arguments{
\item{MFI}{[`MutualForestImpact()`] object.}

\item{candidates}{Vector of variable names that **are candidates to be related to the variables**.}

\item{variables}{Vector of variable names for **which related variables should be searched**.}

\item{p.threshold}{(Default = 0.01) P-value threshold}
}
\value{
A list:
 * `selected`: A list of vectors containing selected candidates for each investigated variable.
 * `p.values`: A list of numeric vectors containing p-values for each candidate's relation to each investigated variable.
}
\description{
Permutation method for MFI.
}
\keyword{internal}
\keyword{varsel}
