% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutualForestImpact.R
\name{MutualForestImpact}
\alias{MutualForestImpact}
\title{Mutual Forest Impact (Corrected Mean Adjusted Agreement).}
\usage{
MutualForestImpact(REL, PERM, variables, candidates)
}
\arguments{
\item{REL}{A [MeanAdjustedAgreement()] object.}

\item{PERM}{A permutated [MeanAdjustedAgreement()] object.}

\item{variables}{Vector of variable names for **which related variables should be searched**.}

\item{candidates}{Vector of variable names that **are candidates to be related to the variables**.}
}
\value{
A `MutualForestImpact` list object.
 * `REL`: The [MeanAdjustedAgreement()] object.
 * `PERM`: The permutated [MeanAdjustedAgreement()] object.
 * `relations`: Matrix of determined relations (rows: investigated variables, columns: candidate variables).
}
\description{
It is recommended to use the simpler [MFI()] method.
}
\seealso{
[MFI()]
}
\keyword{varrel}
