% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutualForestImpact.R
\name{MutualForestImpactVariableSelection}
\alias{MutualForestImpactVariableSelection}
\title{Variable selection for MutualForestImpact.}
\usage{
MutualForestImpactVariableSelection(
  MFI,
  variables,
  candidates,
  p.threshold = 0.01,
  method = c("Janitza", "Permutation")
)
}
\arguments{
\item{MFI}{[MutualForestImpact()] object.}

\item{variables}{Vector of variable names for **which related variables should be searched**.}

\item{candidates}{Vector of variable names that **are candidates to be related to the variables**.}

\item{p.threshold}{(Default = 0.01) P-value threshold}

\item{method}{The method to use. One of: `"Janitza"` or `"Permutation"`.}
}
\value{
A list:
 * `method`: The method used.
 * `selected`: A list of vectors containing selected candidates for each investigated variable.
 * `p.values`: A list of numeric vectors containing p-values for each candidate's relation to each investigated variable.
}
\description{
Variable selection for MutualForestImpact.
}
\keyword{varsel}
