% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutualImpurityReduction.R
\name{MutualImpurityReduction}
\alias{MutualImpurityReduction}
\title{Mutual Impurity Reduction (MIR)}
\usage{
MutualImpurityReduction(REL)
}
\arguments{
\item{REL}{A [MeanAdjustedAgreement()] or [MutualForestImpact()] object.}
}
\value{
A `MutualImpurityReduction` list object.
 * `REL`: The [MeanAdjustedAgreement()] or [MutualForestImpact()] object.
 * `MIR`: Mutual Impurity Reduction
 * `AIR`: Actual Impurity Reduction
}
\description{
Mutual Impurity Reduction (MIR)
}
\keyword{mir}
\keyword{varrel}
