% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutualImpurityReduction.R
\name{MutualImpurityReductionVariableSelection}
\alias{MutualImpurityReductionVariableSelection}
\title{Variable selection for Mutual Impurity Reduction.}
\usage{
MutualImpurityReductionVariableSelection(
  MIR,
  p.threshold = 0.01,
  method = c("Janitza", "Permutation"),
  permutation.num = 100,
  permutation.MeanAdjustedAgreement = NULL
)
}
\arguments{
\item{MIR}{[MutualImpurityReduction()] object.}

\item{p.threshold}{(Default = 0.01) P-value threshold}

\item{method}{The method to use. One of: `"Janitza"` or `"Permutation"`.}

\item{permutation.num}{If method is `"Permutation"`: Number of AIR permutations to determine p-value. (Default: 100)}

\item{permutation.MeanAdjustedAgreement}{If method is `"Permutation"` and `MIR` used a [MeanAdjustedAgreement()] object: A [MeanAdjustedAgreement()] created with a `permutate = TRUE` [RandomForestSurrogates()] object.}
}
\value{
A list:
 * `method`: The method used.
 * `selected`: A list of vectors containing selected candidates for each investigated variable.
 * `p.values`: A list of numeric vectors containing p-values for each candidate's relation to each investigated variable.
}
\description{
Variable selection for Mutual Impurity Reduction.
}
\keyword{mir}
\keyword{varsel}
