% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.relations.R
\name{var.relations}
\alias{var.relations}
\title{Investigate variable relations of a specific variable with mean adjusted agreement}
\usage{
var.relations(
  x = NULL,
  y = NULL,
  num.trees = 500,
  type = "regression",
  s = NULL,
  mtry = NULL,
  min.node.size = 1,
  num.threads = NULL,
  status = NULL,
  save.ranger = FALSE,
  create.forest = is.null(forest),
  forest = NULL,
  save.memory = FALSE,
  case.weights = NULL,
  variables,
  candidates,
  t = 5,
  select.rel = TRUE
)
}
\arguments{
\item{x}{data.frame of predictor variables with variables in
columns and samples in rows (Note: missing values are not allowed)}

\item{y}{vector with values of phenotype variable (Note: will be converted to factor if
classification mode is used). For survival forests this is the time variable.}

\item{num.trees}{number of trees. Default is 500.}

\item{type}{mode of prediction ("regression", "classification" or "survival"). Default is regression.}

\item{s}{predefined number of surrogate splits (it may happen that the actual number of surrogate splits differs in individual nodes). Default is 1 percent of no. of variables.}

\item{mtry}{number of variables to possibly split at in each node. Default is no. of variables^(3/4) ("^3/4") as recommended by (Ishwaran 2011). Also possible is "sqrt" and "0.5" to use the square root or half of the no. of variables.}

\item{min.node.size}{minimal node size. Default is 1.}

\item{num.threads}{number of threads used for determination of relations. Default is number of CPUs available.}

\item{status}{status variable, only applicable to survival data. Use 1 for event and 0 for censoring.}

\item{save.ranger}{set TRUE if ranger object should be saved. Default is that ranger object is not saved (FALSE).}

\item{create.forest}{Default: TRUE if `forest` is NULL, FALSE otherwise. Whether to create or use an existing forest.}

\item{forest}{the random forest that should be analyzed}

\item{save.memory}{Use memory saving (but slower) splitting mode. No effect for survival and GWAS data. Warning: This option slows down the tree growing, use only if you encounter memory problems. (This parameter is transfered to ranger)}

\item{case.weights}{Weights for sampling of training observations. Observations with larger weights will be selected with higher probability in the bootstrap (or subsampled) samples for the trees.}

\item{variables}{variable names (string) for which related variables should be searched for (has to be contained in allvariables)}

\item{candidates}{vector of variable names (strings) that are candidates to be related to the variables (has to be contained in allvariables)}

\item{t}{variable to calculate threshold. Default is 5.}

\item{select.rel}{set False if only relations should be calculated and no related variables should be selected.}
}
\value{
A list containing:
 * `variables`: the variables to which relations are investigated.
 * `surr.res`: a matrix with mean adjusted agreement values with variables in rows and candidates in columns.
 * `threshold`: the threshold used to select related variables.
 * `var`: a list with one vector for each variable containing related variables.
 * `ranger`: ranger object.
}
\description{
This function uses the mean adjusted agreement to select variables that are related to a defined variable using a threshold T.
The parameter t is used to calculate T: t=1 means that every variable with higher probability than "by chance" is identified
as "important". t=2 means the probability has to be twice, etc.
Based on the threshold a vector is created containing the related variables.
}
\examples{
\donttest{
data("SMD_example_data")
x <- SMD_example_data[, 2:ncol(SMD_example_data)]
y <- SMD_example_data[, 1]
set.seed(42)
res <- var.relations(
  x = x,
  y = y,
  s = 10,
  num.trees = 10,
  variables = c("X1", "X7"),
  candidates = colnames(x)[1:100],
  t = 5,
  num.threads = 1
)
res$var
}

}
