% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanAdjustedAgreement.R
\name{MeanAdjustedAgreement}
\alias{MeanAdjustedAgreement}
\title{Investigate variable relations of a specific variable with mean adjusted agreement}
\usage{
MeanAdjustedAgreement(
  RFS,
  t = 5,
  variables = RFS$ranger$forest$independent.variable.names,
  candidates = RFS$ranger$forest$independent.variable.names,
  related = TRUE,
  num.threads = 1
)
}
\arguments{
\item{RFS}{A \code{RandomForestSurrogates} object.}

\item{t}{Used to calculate threshold. (Default: 5)}

\item{variables}{Vector of variable names for \strong{which related variables should be searched}. (Default: All variables used to create the random forest.)}

\item{candidates}{Vector of variable names that \strong{are candidates to be related to the variables}. (Default: All variables used to create the random forest.)}

\item{related}{(Default: \code{TRUE}) Whether related variables should be identified.}

\item{num.threads}{(Default: 1) Number of threads used for determination of relations.}
}
\value{
A \code{MeanAdjustedAgreement} list object:
\itemize{
\item \code{RFS}: The original \code{\link{RandomForestSurrogates}} object.
\item \code{relations}: Matrix with mean adjusted agreement values
\itemize{
\item Rows: \code{variables}.
\item Columns: \code{candidates}.
}
\item \code{threshold}: the threshold used to select related variables.
\item \code{related}: A list of vectors for each \code{variable} containing related \code{candidates}. Only present if \code{related = TRUE} (Default).
}
}
\description{
This function uses the mean adjusted agreement to select variables that are related to a defined variable using a threshold T.
The parameter t is used to calculate T: t=1 means that every variable with higher probability than "by chance" is identified
as "important". t=2 means the probability has to be twice, etc.
Based on the threshold a vector is created containing the related variables.
}
\examples{
\donttest{
data("SMD_example_data")
rfs <- RandomForestSurrogates(
  x = SMD_example_data[, -1]
  y = SMD_example_data[, 1]
  s = 10,
  seed = 42,
  num.trees = 10,
  num.threads = 1
)
maa <- MeanAdjustedAgreement(
  rfs,
  variables = c("X7", "X1"),
  candidates = colnames(SMD_example_data)[2:101],
  t = 5,
  num.threads = 1
)
}

}
\keyword{varrel}
