% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutualForestImpact.R
\name{MFI_VarSel_Janitza}
\alias{MFI_VarSel_Janitza}
\title{Janitza method for MFI.}
\usage{
MFI_VarSel_Janitza(MFI, variables, candidates, p.threshold = 0.01)
}
\arguments{
\item{MFI}{\code{\link[=MutualForestImpact]{MutualForestImpact()}} object.}

\item{variables}{Vector of variable names for \strong{which related variables should be searched}.}

\item{candidates}{Vector of variable names that \strong{are candidates to be related to the variables}.}

\item{p.threshold}{(Default = 0.01) P-value threshold}
}
\value{
A list:
\itemize{
\item \code{selected}: A list of vectors containing selected candidates for each investigated variable.
\item \code{p.values}: A list of numeric vectors containing p-values for each candidate's relation to each investigated variable.
}
}
\description{
Janitza method for MFI.
}
\keyword{internal}
\keyword{varsel}
