% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutualForestImpact.R
\name{MutualForestImpactVariableSelection}
\alias{MutualForestImpactVariableSelection}
\title{Variable selection for MutualForestImpact.}
\usage{
MutualForestImpactVariableSelection(
  MFI,
  variables,
  candidates,
  p.threshold = 0.01,
  method = c("Janitza", "Permutation")
)
}
\arguments{
\item{MFI}{\code{\link[=MutualForestImpact]{MutualForestImpact()}} object.}

\item{variables}{Vector of variable names for \strong{which related variables should be searched}.}

\item{candidates}{Vector of variable names that \strong{are candidates to be related to the variables}.}

\item{p.threshold}{(Default = 0.01) P-value threshold}

\item{method}{The method to use. One of: \code{"Janitza"} or \code{"Permutation"}.}
}
\value{
A list:
\itemize{
\item \code{method}: The method used.
\item \code{selected}: A list of vectors containing selected candidates for each investigated variable.
\item \code{p.values}: A list of numeric vectors containing p-values for each candidate's relation to each investigated variable.
}
}
\description{
Variable selection for MutualForestImpact.
}
\keyword{varsel}
