% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutualImpurityReduction.R
\name{MutualImpurityReduction}
\alias{MutualImpurityReduction}
\title{MutualForestImpact.}
\usage{
MutualImpurityReduction(REL)
}
\arguments{
\item{REL}{A \code{\link{MeanAdjustedAgreement}} or \code{\link{MutualForestImpract}} object.}
}
\value{
A \code{\link{MutualImpurityReduction}} list object.
\itemize{
\item \code{REL}: The \code{\link{MeanAdjustedAgreement}} or \code{\link{MutualForestImpact}} object.
\item \code{MIR}: TODO
\item \code{AIR}: TODO
}
}
\description{
MutualForestImpact.
}
\keyword{varsel}
