% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutualImpurityReduction.R
\name{MutualImpurityReduction}
\alias{MutualImpurityReduction}
\title{Mutual Impurity Reduction (MIR)}
\usage{
MutualImpurityReduction(REL)
}
\arguments{
\item{REL}{A \code{\link[=MeanAdjustedAgreement]{MeanAdjustedAgreement()}} or \code{\link[=MutualForestImpact]{MutualForestImpact()}} object.}
}
\value{
A \code{MutualImpurityReduction} list object.
\itemize{
\item \code{REL}: The \code{\link[=MeanAdjustedAgreement]{MeanAdjustedAgreement()}} or \code{\link[=MutualForestImpact]{MutualForestImpact()}} object.
\item \code{MIR}: Mutual Impurity Reduction
\item \code{AIR}: Actual Impurity Reduction
}
}
\description{
Mutual Impurity Reduction (MIR)
}
\keyword{mir}
\keyword{varrel}
