% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutualImpurityReduction.R
\name{MutualImpurityReductionVariableSelection}
\alias{MutualImpurityReductionVariableSelection}
\title{Variable selection for MutualImpurityReduction.}
\usage{
MutualImpurityReductionVariableSelection(
  MIR,
  p.threshold = 0.01,
  method = c("Janitza", "Permutation"),
  permutation.num = 100
)
}
\arguments{
\item{MIR}{\code{\link[=MutualImpurityReduction]{MutualImpurityReduction()}} object.}

\item{p.threshold}{(Default = 0.01) P-value threshold}

\item{method}{The method to use. One of: \code{"Janitza"} or \code{"Permutation"}.}

\item{permutation.num}{(If method is \code{"Permutation"}) Number of AIR permutations to determine p-value. (Default: 100)}
}
\value{
A list:
\itemize{
\item \code{method}: The method used.
\item \code{selected}: A list of vectors containing selected candidates for each investigated variable.
\item \code{p.values}: A list of numeric vectors containing p-values for each candidate's relation to each investigated variable.
}
}
\description{
Variable selection for MutualImpurityReduction.
}
\keyword{varsel}
