% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForestSurrogates.R
\name{RandomForestSurrogates}
\alias{RandomForestSurrogates}
\title{Create a random forest with surrogates.}
\usage{
RandomForestSurrogates(
  x = NULL,
  y = NULL,
  s.pct = 0.01,
  s = ceiling(ncol(x) * ifelse(s.pct > 0 && s.pct <= 1, s.pct, 0.01)),
  mtry = c("^3/4", "sqrt", "0.5"),
  type = c("regression", "classification", "survival"),
  status = NULL,
  num.trees = 500,
  num.threads = parallel::detectCores(),
  min.node.size = 1,
  permutate = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{Predictor data (independent variables), alternative interface to data with formula or dependent.variable.name.}

\item{y}{Response vector (dependent variable), alternative interface to data with formula or dependent.variable.name. For survival use a \code{Surv()} object or a matrix with time and status.}

\item{s.pct, s}{Number of surrogate splits.
This can be defined either by setting \code{s.pct} to a number between
0 and 1, or providing an exact value for \code{s}.
\itemize{
\item \code{s.pct}: Percentage of variables to use for \code{s}. (Default: 0.01)
\item \code{s}:  Number of surrogate splits. (Default: Number of variables
multiplied by \code{s.pct}, which defaults to 0.01; If \code{s.pct} is
less than or equal to zero, or greater than 1: 0.01 is used instead.)
}}

\item{mtry}{Number of variables to possibly split at in each node.
Default is the (rounded down) number of variables to the power
of three quarters (Ishwaran, 2011).
Alternatively, a single argument function returning an integer,
given the number of independent variables.}

\item{type}{The type of random forest to create with ranger.
One of "regression" (Default), "classification" or "survival".}

\item{num.trees}{Number of trees.}

\item{num.threads}{Number of threads. Default is number of CPUs available.}

\item{min.node.size}{Minimal node size to split at. (Default: 1)}

\item{permutate}{Enable to permutate \code{x} for \code{\link{MutualForestImpact}} (Default: FALSE).}

\item{seed}{RNG seed. It is strongly recommended that you set this value.}

\item{...}{Other params passed on to \code{\link[ranger:ranger]{ranger::ranger()}}.}
}
\value{
A \code{RandomForestSurrogates} S3 object.
\itemize{
\item \code{trees}: List of all trees with surrogate analysis. (Class: \code{SurrogateTrees}, \code{LayerTrees}, \code{RangerTrees})
\item \code{ranger}: \link[ranger:ranger]{ranger::ranger} model used to obtain the trees.
\item \code{s} = s: The \code{s} parameter.
}
}
\description{
Create a random forest with surrogates.
}
\keyword{prep}
