% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrmindep.R
\name{surrmindep}
\alias{surrmindep}
\title{Execute surrogate minimal depth variable importance}
\usage{
surrmindep(forest, s.l = count.surrogates(forest$trees)$s.l)
}
\arguments{
\item{forest}{a list containing \code{allvariables} and \code{trees}.
Allvariables is a vector of all variable names in the original data set (strings).
Trees is a list of trees that was generated by \code{\link[=getTreeranger]{getTreeranger()}},
\code{\link[=addLayer]{addLayer()}}, and \code{\link[=addSurrogates]{addSurrogates()}}.}

\item{s.l}{Number of average surrogate variables in the respective layers.
(use \code{\link[=count.surrogates]{count.surrogates()}} function to get it)}
}
\value{
List with the following components:
\itemize{
\item \code{depth}: mean surrogate minimal depth for each variable
\item \code{selected}: variables has been selected (1) or not (0),
\item \code{threshold}: the threshold that is used for the selection
}
}
\description{
This function determines the surrogate minimal depth of variables from a
forest that is created by \code{\link[=getTreeranger]{getTreeranger()}}, \code{\link[=addLayer]{addLayer()}} and \code{\link[=addSurrogates]{addSurrogates()}} functions.
}
